expand scope of arxiv identifier matcher
The simple part of this is allowing 'arxiv:' in addition to 'arXiv:'.

The more complex second part is to conservatively match "old" (pre-2008)
style identifiers which do not have a prefix. The conservative matching
is because there is less confidence that a string is actually an arxiv
identifier without the prefix. Explicit collection prefixes are included
(for those that existed pre-2008), internal whitespace is not allowed,
and the identifier must be separated from other alphabetic strings.